Align namespaces for ISiloBuilder, IClientBuilder, and IServiceCollection extension methods
