+++ src/AWS/Orleans.Reminders.DynamoDB/DynamoDBServiceCollectionReminderExtensions.cs
@@ -1,0 +1,34 @@
+using Microsoft.Extensions.DependencyInjection;
+using Orleans;
+using Orleans.Configuration;
+using Orleans.Reminders.DynamoDB;
+using System;
+
+namespace Orleans.Hosting
+{
+    /// <summary>
+    /// <see cref="IServiceCollection"/> extensions.
+    /// </summary>
+    public static class DynamoDBServiceCollectionReminderExtensions
+    {
+        /// <summary>
+        /// Adds reminder storage backed by Amazon DynamoDB.
+        /// </summary>
+        /// <param name="services">
+        /// The service collection.
+        /// </param>
+        /// <param name="configure">
+        /// The delegate used to configure the reminder store.
+        /// </param>
+        /// <returns>
+        /// The provided <see cref="IServiceCollection"/>, for chaining.
+        /// </returns>
+        public static IServiceCollection UseDynamoDBReminderService(this IServiceCollection services, Action<DynamoDBReminderStorageOptions> configure)
+        {
+            services.AddSingleton<IReminderTable, DynamoDBReminderTable>();
+            services.Configure<DynamoDBReminderStorageOptions>(configure);
+            services.ConfigureFormatter<DynamoDBReminderStorageOptions>();
+            return services;
+        }
+    }
+}
