+++ src/AWS/Orleans.Reminders.DynamoDB/DynamoDBSiloBuilderReminderExtensions.cs
@@ -8,7 +8,7 @@ namespace Orleans.Hosting
     /// <summary>
     /// Silo host builder extensions.
     /// </summary>
-    public static class SiloBuilderReminderExtensions
+    public static class DynamoDBSiloBuilderReminderExtensions
     {
         /// <summary>
         /// Adds reminder storage backed by Amazon DynamoDB.
@@ -26,26 +26,6 @@ namespace Orleans.Hosting
         {
             builder.ConfigureServices(services => services.UseDynamoDBReminderService(configure));
             return builder;
-        }
-
-        /// <summary>
-        /// Adds reminder storage backed by Amazon DynamoDB.
-        /// </summary>
-        /// <param name="services">
-        /// The service collection.
-        /// </param>
-        /// <param name="configure">
-        /// The delegate used to configure the reminder store.
-        /// </param>
-        /// <returns>
-        /// The provided <see cref="IServiceCollection"/>, for chaining.
-        /// </returns>
-        public static IServiceCollection UseDynamoDBReminderService(this IServiceCollection services, Action<DynamoDBReminderStorageOptions> configure)
-        {
-            services.AddSingleton<IReminderTable, DynamoDBReminderTable>();
-            services.Configure<DynamoDBReminderStorageOptions>(configure);
-            services.ConfigureFormatter<DynamoDBReminderStorageOptions>();
-            return services;
         }
     }
 }
