+++ src/Azure/Orleans.Reminders.AzureStorage/AzureStorageReminderServiceCollectionExtensions.cs
@@ -1,6 +1,7 @@
 using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using Orleans;
 using Orleans.Configuration;
 using Orleans.Reminders.AzureStorage;
 using Orleans.Runtime.ReminderService;
@@ -8,64 +9,10 @@ using Orleans.Runtime.ReminderService;
 namespace Orleans.Hosting
 {
     /// <summary>
-    /// Silo host builder extensions.
+    /// <see cref="IServiceCollection"/> extensions.
     /// </summary>
-    public static class SiloBuilderReminderExtensions
+    public static class AzureStorageReminderServiceCollectionExtensions
     {
-        /// <summary>
-        /// Adds reminder storage backed by Azure Table Storage.
-        /// </summary>
-        /// <param name="builder">
-        /// The builder.
-        /// </param>
-        /// <param name="configure">
-        /// The delegate used to configure the reminder store.
-        /// </param>
-        /// <returns>
-        /// The provided <see cref="ISiloBuilder"/>, for chaining.
-        /// </returns>
-        public static ISiloBuilder UseAzureTableReminderService(this ISiloBuilder builder, Action<AzureTableReminderStorageOptions> configure)
-        {
-            builder.ConfigureServices(services => services.UseAzureTableReminderService(configure));
-            return builder;
-        }
-
-        /// <summary>
-        /// Adds reminder storage backed by Azure Table Storage.
-        /// </summary>
-        /// <param name="builder">
-        /// The builder.
-        /// </param>
-        /// <param name="configureOptions">
-        /// The configuration delegate.
-        /// </param>
-        /// <returns>
-        /// The provided <see cref="ISiloBuilder"/>, for chaining.
-        /// </returns>
-        public static ISiloBuilder UseAzureTableReminderService(this ISiloBuilder builder, Action<OptionsBuilder<AzureTableReminderStorageOptions>> configureOptions)
-        {
-            builder.ConfigureServices(services => services.UseAzureTableReminderService(configureOptions));
-            return builder;
-        }
-
-        /// <summary>
-        /// Adds reminder storage backed by Azure Table Storage.
-        /// </summary>
-        /// <param name="builder">
-        /// The builder.
-        /// </param>
-        /// <param name="connectionString">
-        /// The storage connection string.
-        /// </param>
-        /// <returns>
-        /// The provided <see cref="ISiloBuilder"/>, for chaining.
-        /// </returns>
-        public static ISiloBuilder UseAzureTableReminderService(this ISiloBuilder builder, string connectionString)
-        {
-            builder.UseAzureTableReminderService(options => options.ConfigureTableServiceClient(connectionString));
-            return builder;
-        }
-
         /// <summary>
         /// Adds reminder storage backed by Azure Table Storage.
         /// </summary>
