+++ src/Azure/Orleans.Reminders.AzureStorage/AzureStorageReminderSiloBuilderReminderExtensions.cs
@@ -1,0 +1,67 @@
+using System;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Options;
+using Orleans.Reminders.AzureStorage;
+
+namespace Orleans.Hosting
+{
+    /// <summary>
+    /// Silo host builder extensions.
+    /// </summary>
+    public static class AzureStorageReminderSiloBuilderExtensions
+    {
+        /// <summary>
+        /// Adds reminder storage backed by Azure Table Storage.
+        /// </summary>
+        /// <param name="builder">
+        /// The builder.
+        /// </param>
+        /// <param name="configure">
+        /// The delegate used to configure the reminder store.
+        /// </param>
+        /// <returns>
+        /// The provided <see cref="ISiloBuilder"/>, for chaining.
+        /// </returns>
+        public static ISiloBuilder UseAzureTableReminderService(this ISiloBuilder builder, Action<AzureTableReminderStorageOptions> configure)
+        {
+            builder.ConfigureServices(services => services.UseAzureTableReminderService(configure));
+            return builder;
+        }
+
+        /// <summary>
+        /// Adds reminder storage backed by Azure Table Storage.
+        /// </summary>
+        /// <param name="builder">
+        /// The builder.
+        /// </param>
+        /// <param name="configureOptions">
+        /// The configuration delegate.
+        /// </param>
+        /// <returns>
+        /// The provided <see cref="ISiloBuilder"/>, for chaining.
+        /// </returns>
+        public static ISiloBuilder UseAzureTableReminderService(this ISiloBuilder builder, Action<OptionsBuilder<AzureTableReminderStorageOptions>> configureOptions)
+        {
+            builder.ConfigureServices(services => services.UseAzureTableReminderService(configureOptions));
+            return builder;
+        }
+
+        /// <summary>
+        /// Adds reminder storage backed by Azure Table Storage.
+        /// </summary>
+        /// <param name="builder">
+        /// The builder.
+        /// </param>
+        /// <param name="connectionString">
+        /// The storage connection string.
+        /// </param>
+        /// <returns>
+        /// The provided <see cref="ISiloBuilder"/>, for chaining.
+        /// </returns>
+        public static ISiloBuilder UseAzureTableReminderService(this ISiloBuilder builder, string connectionString)
+        {
+            builder.UseAzureTableReminderService(options => options.ConfigureTableServiceClient(connectionString));
+            return builder;
+        }
+    }
+}
