+++ src/Orleans.CodeGenerator.MSBuild/buildMultiTargeting/Microsoft.Orleans.CodeGenerator.MSBuild.targets
@@ -1,3 +1,3 @@
 <Project>
-  <Import Project="..\build\Orleans.CodeGenerator.MSBuild.targets" />
+  <Import Project="..\build\Microsoft.Orleans.CodeGenerator.MSBuild.targets" />
 </Project>
