+++ src/Orleans.CodeGenerator/buildMultiTargeting/Microsoft.Orleans.CodeGenerator.props
@@ -1,3 +1,3 @@
 <Project>
-  <Import Project="..\build\Orleans.CodeGenerator.props" />
+  <Import Project="..\build\Microsoft.Orleans.CodeGenerator.props" />
 </Project>
