+++ src/Orleans.Connections.Security/Hosting/HostingExtensions.IClientBuilder.cs
@@ -3,9 +3,9 @@ using System.Security.Cryptography.X509Certificates;
 using Orleans.Configuration;
 using Orleans.Connections.Security;
 
-namespace Orleans
+namespace Orleans.Hosting
 {
-    public static class OrleansConnectionSecurityHostingExtensions
+    public static partial class OrleansConnectionSecurityHostingExtensions
     {
         /// <summary>
         /// Configures TLS.
