+++ src/Orleans.Sdk/buildMultiTargeting/Microsoft.Orleans.Sdk.targets
@@ -1,0 +1,3 @@
+<Project>
+  <Import Project="..\build\Microsoft.Orleans.Sdk.targets" />
+</Project>
