+++ src/Orleans.TestingHost/IClientBuilderConfigurator.cs
@@ -1,4 +1,5 @@
 using Microsoft.Extensions.Configuration;
+using Orleans.Hosting;
 
 namespace Orleans.TestingHost
 {
