+++ src/Orleans.TestingHost/InMemoryTransport/InMemoryTransportListenerFactory.cs
@@ -2,20 +2,17 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Net;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Channels;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Connections;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Orleans.Configuration;
 using Orleans.Hosting;
 using Orleans.Internal;
 using Orleans.Networking.Shared;
 using Orleans.Runtime;
 using Orleans.Runtime.Messaging;
-using Orleans.Timers.Internal;
 
 namespace Orleans.TestingHost.InMemoryTransport;
