+++ src/Orleans.Transactions/Hosting/SiloBuilderExtensions.cs
@@ -1,6 +1,4 @@
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-using Orleans.Runtime;
 using Orleans.Transactions.Abstractions;
 using Orleans.Transactions;
 
@@ -19,24 +17,6 @@ namespace Orleans.Hosting
             return builder.ConfigureServices(services => services.UseTransactions(withStatisticsReporter))
                           .AddGrainExtension<ITransactionManagerExtension, TransactionManagerExtension>()
                           .AddGrainExtension<ITransactionalResourceExtension, TransactionalResourceExtension>();
-        }
-
-        internal static IServiceCollection UseTransactions(this IServiceCollection services, bool withReporter)
-        {
-            services.TryAddSingleton<IClock,Clock>();
-            services.TryAddSingleton<ITransactionAgentStatistics, TransactionAgentStatistics>();
-            services.TryAddSingleton<ITransactionOverloadDetector,TransactionOverloadDetector>();
-            services.AddSingleton<ITransactionAgent, TransactionAgent>();
-            services.TryAddSingleton(typeof(ITransactionDataCopier<>), typeof(DefaultTransactionDataCopier<>));
-            services.AddSingleton<IAttributeToFactoryMapper<TransactionalStateAttribute>, TransactionalStateAttributeMapper>();
-            services.TryAddTransient<ITransactionalStateFactory, TransactionalStateFactory>();
-            services.AddSingleton<IAttributeToFactoryMapper<TransactionCommitterAttribute>, TransactionCommitterAttributeMapper>();
-            services.TryAddTransient<ITransactionCommitterFactory, TransactionCommitterFactory>();
-            services.TryAddTransient<INamedTransactionalStateStorageFactory, NamedTransactionalStateStorageFactory>();
-            services.AddTransient(typeof(ITransactionalState<>), typeof(TransactionalState<>));
-            if (withReporter)
-                services.AddSingleton<ILifecycleParticipant<ISiloLifecycle>, TransactionAgentStatisticsReporter>();
-            return services;
         }
     }
 }
