+++ test/Directory.Build.props
@@ -21,4 +21,5 @@
   <ItemGroup>
     <None Include="*.xunit.runner.json" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
+
 </Project>
